Add support for storing multiple auth tokens in fixie auth (#102)
This adds support for multiple auth tokens being stored in the Fixie CLI
config. The config contains a default API token as well as a number of
URL-specific tokens:

- When looking for a token we first check if there's a token specific to
the current Fixie URL and fall back to the default token
- When updating the token we add the token to the host-specific tokens
list _as well as the default token_. This means that for any
previously-unseen URL, the last token recorded by `fixie auth` will be
used.