Add response chunking and auth token verification to `GET /` (#100)
This adds response chunking and auth token verification to `GET /`. For
auth token verification, in addition to verifying that the response has
a valid Fixie request token we also verify that the agent ID in the
token is the expected ID.

To avoid requiring that the agent ID be embedded in the _code_, the
`fixie serve` and `fixie deploy` commands set the
`FIXIE_ALLOWED_AGENT_ID` environment variable. The latter loads an
`.env` file in the bootstrapping code. This environment variable is
optional to avoid it being a breaking change -- we can make it required
in the future. To avoid confusion between multiple uses of `agent_id` in
`CodeShotAgent` this moves the refresh-on-startup logic entirely out of
the agent implementation and into the agent loader where `fixie serve`
can continue to use it.

Finally, to allow agents to be deployed against non-prod Fixie
environments, this uses the new `.env` mechanism to set `FIXIE_API_URL`
to match the environment that the agent is being deployed into.