Improved error handling in the voice client. (#210)
This PR makes a few minor changes to improve error handling and
debugging.

- We explicitly track the initializing state of the VoiceSession as it
goes from created -> connecting -> connected -> audio enabled ->
started. This is going to make it easier for applications to observe the
state of the session and take appropriate action when errors or timeouts
occur.
- The onError handler carries a VoiceSessionError object with
information about the error. This is a simple message for now, but can
be expanded later.
- Added a few additional places where .onError is called.

The more substantive (perhaps) change is that .warmup is no longer
called in the ctor or when the socket closes. I would much rather this
be done explicitly by application code, rather than the VoiceSession
trying to be too clever. Constructors are not a good place to be doing
actions like opening socket connections, anyway, because it's much
harder to deal with errors that happen during the ctor. By having
.warmup() be called explicitly, we can create the VoiceSession, set up
the error handling logic, call .warmup, and deal gracefully with
connection errors that might occur.

These changes allow us to have more graceful error handling in the Santa
app.