Working on auth.test.ts.
For some reason jest hangs when running this test if I define
the `mockFetch` function, even if it's not used. My guess here
is that ts-jest or something else in the Jest pipeline is choking
on the code itself, not trying to execute the code. What is odd
is that it works in the fixie-common tests.