Refactoring of Fixie JS SDK. (#216)
This PR refactors the SDK into three packages:

* `@fixieai/fixie-common` - Base package for both NodeJS and browser
environments, containing common code for both. Not intended to be used
directly.
* `fixie` - NodeJS package, based on top of `@fixieai/fixie-common`,
containing code for the CLI, agent deployment, etc.
* `fixie-web`- Browser package, based on top of `@fixieai/fixie-common`,
containing code for browser environments, including the Voice Session
API.

The goals here are to make it possible to cleanly separate code for the
browser and NodeJS environments, manage dependencies in a sane way
between the two environments, and avoid code duplication by having a
common environment-neutral package.

This PR also adds unit tests for several of the modules to provide some
baseline structure for testing and CI going forward.

For current users of the Fixie package in NodeJS environments,
everything should work as it currently does. The CLI works identically
as well.

Browser-based users will need to migrate to the new `fixie-web` package
for voice and React hooks, etc.