Fix problem with 'main' not being defined for the 'fixie' package. (#231)
This fixes a problem where the 'fixie' package could not be imported by
external code because the 'main' was not defined in package.json. It
also fixes a problem with the auth tests when the user running the tests
has a FIXIE_API_KEY defined, and adds a trivial example of using the
Fixie SDK from a Node client.